web/requirements.txt
Flask==3.1.1
gunicorn==23.0.0
python-dotenv==1.0.1
Pillow==12.2.0
flask-compress==1.24
Flask-WTF==1.2.2
Flask-Limiter==3.12
markdown==3.8
bleach==6.2.0
pytest==8.4.0
